add document service
